moved helper functions to seperate files in helper folder
