+++ app/assets/main/react_components/helpers/UserActionsHelper.js
@@ -1,0 +1,31 @@
+export const formatedUserActions = (inVal) => {
+  return inVal.map((userActions) => ({
+    ...userActions,
+    id: userActions.id.toString(),
+  }));
+};
+
+export const acceptedUserActions = (inVal) => {
+  return formatedUserActions(inVal)
+    .filter((action) => action.status !== true)
+    .map((action) => ({ ...action }));
+};
+
+export const columnUserActions = (
+  acceptedList,
+  doneList,
+  climateActionsText
+) => {
+  return {
+    [1]: {
+      id: "Accepted",
+      name: climateActionsText.my_actions,
+      items: acceptedList,
+    },
+    [2]: {
+      id: "Performed",
+      name: climateActionsText.achievements,
+      items: doneList,
+    },
+  };
+};
